deploy: return accidentally deleted default profile picture and add credentials: 'include' to requests that set cookies

+++ api/src/web/file-manipulation/delete-files.ts
@@ -1,7 +1,11 @@
 import fs from 'fs';
 import path from 'path';
 
-import { POST_IMAGES_LOCATION, PROFILE_PICS_LOCATION } from 'config';
+import {
+  POST_IMAGES_LOCATION,
+  PROFILE_PICS_LOCATION,
+  DEFAULT_PROFILE_PICTURE,
+} from 'config';
 
 export const deleteFile = (filePath: string) => {
   fs.unlink(filePath, (err) => err && console.error(err));
@@ -11,6 +15,7 @@ export const deletePostImage = (fileName: string) => {
   deleteFile(path.join(POST_IMAGES_LOCATION, fileName));
 };
 
-export const deleteProfilePic = (fileName: string) => {
+export const deleteProfilePic = (fileName: string, force: boolean = false) => {
+  if (fileName === DEFAULT_PROFILE_PICTURE && !force) return;
   deleteFile(path.join(PROFILE_PICS_LOCATION, fileName));
 };
